Add user logins

+++ src/Core/Brewdude.Domain/Entities/Beer.cs
@@ -8,7 +8,7 @@ namespace Brewdude.Domain.Entities
         public string Name { get; set; }
         public string Description { get; set; }
         public BeerStyle BeerStyle { get; set; }
-        public byte Ibu { get; set; }
+        public int Ibu { get; set; }
         public double Abv { get; set; }
         public int? BreweryId { get; set; }
         public virtual Brewery Brewery { get; set; }
